% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tab}
\alias{tab}
\title{Tabulate two variables}
\usage{
tab(x, y)
}
\arguments{
\item{x}{A character.}

\item{y}{A character.}
}
\value{
The summary of counts of \code{x} and \code{y}.
}
\description{
Tabulate two variables
}
\examples{
color <- c("yellow", "yellow", "white")
flower <- c("sunflower", "rose", "daisy")
tab(color, flower)
}
